Pushing down the VisitorProvider into the lint and format functions resulted in a performance loss and lots of duplicate log lines regarding the initialization of the VisitorProvider. Also the "ktlint" task in the build pipeline was no longer usable as the relevant lines were almost impossible to spot between all duplicate log lines.
Pushing down the VisitorProvider is however still an improvement that is to be kept. The initialization of the VisitorProvider is now extracted to the VisitorProviderFactory. This factory only initializes the VisitorProvider whenever it is called for a new combination of rules. Otherwise, the previously created instance is reused. As a result the logging is only printed for unique combinations of rules.

Als the RuleReference and QualifiedRuleId are extracted from the VisitorProvider so that they can e=be reused in the VisitorProviderFactory.
